rtp::packetizer

Trait Depacketizer

pub trait Depacketizer {
    // Required methods
    fn depacketize(&mut self, b: &Bytes) -> Result<Bytes, Error>;
    fn is_partition_head(&self, payload: &Bytes) -> bool;
    fn is_partition_tail(&self, marker: bool, payload: &Bytes) -> bool;
}

Depacketizer depacketizes a RTP payload, removing any RTP specific data from the payload

Required Methods

fn depacketize(&mut self, b: &Bytes) -> Result<Bytes, Error>

fn is_partition_head(&self, payload: &Bytes) -> bool

Checks if the packet is at the beginning of a partition. This should return false if the result could not be determined, in which case the caller will detect timestamp discontinuities.

fn is_partition_tail(&self, marker: bool, payload: &Bytes) -> bool

Checks if the packet is at the end of a partition. This should return false if the result could not be determined.